refactor: 리덕스 파일 src에 옮김+필요없는 것들 삭제...
