--- pages/todoRoom/todoRoomImports.js
@@ -1,5 +1,0 @@
-import Todo_btn from "../../src/components/todoRoom/buttons/Todo_btn"
-import TodoCard from "../../src/components/todoRoom/buttons/todoCard"
-import Done_btn from "../../src/components/todoRoom/buttons/Done_btn"
-
-export default {Todo_btn, Doing_btn: TodoCard, Done_btn}
