--- src/features/todoRoom/buttons/CommonButton.tsx
@@ -1,7 +1,0 @@
-import styled from 'styled-components';
-
-const button = styled.button`
-
-  
-  
-`
