--- src/features/todoRoom/buttons/todoCardMethod.ts
@@ -1,19 +1,0 @@
-// import react from 'react';
-// import {useSelector, useDispatch} from 'react-redux';
-// import { addDoing } from '../../../redux/todoRoom/Doings';
-//
-//
-// const Todo_Doing = () => {
-//   console.log("doing")
-// }
-// const Doing_Done = () => {
-//
-// }
-// const Done_Doing = () => {
-//
-// }
-// const Doing_Todo = () => {
-//
-// }
-//
-// export {Todo_Doing, Doing_Done, Done_Doing, Doing_Todo}
