--- src/features/todoRoom/redux_todoRoom/initState.ts
@@ -1,7 +1,0 @@
-import { nanoid } from '@reduxjs/toolkit';
-import TodoType from './todoType';
-
-const initState:TodoType[]= [
-
-]
-export default initState;
