+++ src/redux/todoRoom/Dones.ts
@@ -1,10 +1,10 @@
 import { createSlice, nanoid, PayloadAction } from '@reduxjs/toolkit';
 import { Interface } from 'readline';
-import tempData from './initState';
+import initState from './initState';
 
 const DoneSlice=createSlice({
   name: 'done',
-  initialState:tempData,
+  initialState:initState,
   reducers: {
     addDone: {
       reducer: (state, action: PayloadAction) => {
